entry/kvm: handle TIF_NEED_RESCHED_LAZY
Executing in xfer_to_guest_mode_work() we are free of kernel
entanglements that cannot span preemption.

So, handle TIF_NEED_RESCHED_LAZY alongside TIF_NEED_RESCHED.

Also, while we at it, remove the explicit check for need_resched()
in the exit condition as that is already covered in the loop
condition.

Originally-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ankur Arora <ankur.a.arora@oracle.com>